Separated memory transfer timing information from function timing information and changed maxima to be of type int

+++ include/sma_cuda_func.cuh
@@ -1,4 +1,4 @@
 #pragma once
 
-__global__ void find_minima_cuda(float* raw_data, float* maxima, float* minima);
-__global__ void find_peaks_cuda(float* window_av, float* maxima, int* NUM_WINDOWS);
+__global__ void find_minima_cuda(float* raw_data, int* maxima, float* minima);
+__global__ void find_peaks_cuda(float* window_av, int* maxima, int* NUM_WINDOWS);
